requirements.txt
python-telegram-bot==20.7
yt-dlp>=2024.12.13
